requirements.txt
# Python dependencies
fastapi
uvicorn
asyncpg
pydantic
google-generativeai>=0.5.0
pydantic-settings